move app components out of directories
